GameDetail
This is the code for a GameDetailSceneViewModel class in Swift. The class conforms to the GameDetailSceneViewModelProtocol and GameDetailSceneViewModelDelegate protocols. The class has a delegate property, which is an optional GameDetailSceneViewModelDelegate. The class also has a private game property, which is an optional GameDetailModel. The class has methods for fetching a game's details by ID, getting various properties of a game such as its title, publisher, rating, and release date, as well as methods for handling whether a game is a favorite and liking/unliking a game. The class also has a method that utilizes a RawgClient to retrieve the game's details and a method that uses the Globals class to format the game's release date.